Move the client calls to the runtime pkg
The runtime pkg is going to be responsible for handeling
all the API and client calls.
