+++ pkg/apb/watch_pod.go
@@ -5,6 +5,8 @@ import (
 	"strings"
 	"time"
 
+	"github.com/openshift/ansible-service-broker/pkg/runtime"
+
 	logging "github.com/op/go-logging"
 )
 
@@ -20,7 +22,7 @@ func watchPod(podName string, namespace string, log *logging.Logger) (string, er
 
 	for r := 1; r <= apbWatchRetries; r++ {
 		log.Info("Watch pod [ %s ] tick %d", podName, r)
-		output, err := RunCommand(
+		output, err := runtime.RunCommand(
 			"oc", "get", "pod", "--no-headers=true", "--namespace="+namespace, podName)
 
 		outStr := string(output)
